Split Subgraph API functions into separate files
PiperOrigin-RevId: 314585364
