+++ BUILD.bazel
@@ -74,6 +74,25 @@ OPERATOR_SRCS = [
     "src/operators/unpooling-nhwc.c",
 ]
 
+SUBGRAPH_SRCS = [
+    "src/subgraph/add2.c",
+    "src/subgraph/argmax-pooling-2d.c",
+    "src/subgraph/average-pooling-2d.c",
+    "src/subgraph/clamp.c",
+    "src/subgraph/convolution-2d.c",
+    "src/subgraph/deconvolution-2d.c",
+    "src/subgraph/depthwise-convolution-2d.c",
+    "src/subgraph/fully-connected.c",
+    "src/subgraph/hardswish.c",
+    "src/subgraph/max-pooling-2d.c",
+    "src/subgraph/multiply2.c",
+    "src/subgraph/prelu.c",
+    "src/subgraph/sigmoid.c",
+    "src/subgraph/softmax.c",
+    "src/subgraph/static-constant-pad.c",
+    "src/subgraph/unpooling-2d.c",
+]
+
 TABLE_SRCS = [
     "src/tables/exp2-k-over-64.c",
     "src/tables/exp2-k-over-2048.c",
@@ -2653,7 +2672,7 @@ xnnpack_cc_library(
         "src/runtime.c",
         "src/subgraph.c",
         "src/tensor.c",
-    ],
+    ] + SUBGRAPH_SRCS,
     hdrs = ["include/xnnpack.h"],
     copts = LOGGING_COPTS + [
         "-Isrc",
@@ -2691,7 +2710,7 @@ xnnpack_cc_library(
         "src/runtime.c",
         "src/subgraph.c",
         "src/tensor.c",
-    ],
+    ] + SUBGRAPH_SRCS,
     hdrs = ["include/xnnpack.h"],
     copts = LOGGING_COPTS + [
         "-Isrc",
